pyproject.toml
[tool.poetry]
name = "jellyfish-case-backend"
version = "0.1.0"
description = ""
authors = ["JB Guerin <jeanbaptisteguerin1@outlook.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.101.0"
pydantic = "^1.8.2"
sqlmodel = "^0.0.8"
pyjwt = "^2.8.0"
schedule = "^1.2.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^22.10.0"
pytest-watch = "^4.2.0"
pylint = "^2.17.4"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = "integration"
python_files = "tests.py test_*.py *_test.py"
addopts = "-W ignore::DeprecationWarning"

[tool.black]
line-length = 120

[tool.pylint.master]
ignore = ".venv"
unsafe-load-any-extension = "no"
max-line-length = 120

[tool.pylint.messages_control]
disable = "C0103,C0114,C0115,C0116,C0415,E1101,R0901,R0902,R0903,R0904,R0911,R0912,R0913,R0914,R0915,R0916,I1101,W1203,W0511"

[tool.pylint.similarities]
ignore-imports = "yes"
ignore-signatures = "yes"

[tool.mypy]
ignore_missing_imports = true
